JPS 1.0

+++ include/planner/planner.h
@@ -6,10 +6,10 @@
 #include<vector>
 using namespace std;
 struct Point{
-    int x,y,z;
-    int F,G,H;
+    float x,y,z;
+    float F,G,H;
     Point* parent;
-    Point(int _x,int _y,int _z):x(_x),y(_y),z(_z),F(0),G(0),H(0),parent(NULL){};
+    Point(float _x,float _y,float _z):x(_x),y(_y),z(_z),F(0),G(0),H(0),parent(NULL){};
     
 };
  class Abstract_planner{
